Integrate WI identity to the nomad-consul-secure scenario
This PR integrates the Consul WI configuration into the main scenario.

The existing WI scenario was using the older provisioning scripts and were using user-v2 networking.

The WI will be enabled when the Nomad version is >=1.8.0 and can be disabled by passing `NOMAD_CONSUL_WI` environment variable to `shikari create command.

Update the README with the details.